comment: 사용하지 않는 클래스 임시 주석 처리
Related to: #22
