+++ src/main/java/com/example/omg_project/domain/trip/dto/TripLocationDto.java
@@ -1,15 +1,15 @@
-package com.example.omg_project.domain.trip.dto;
-
-import lombok.*;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class TripLocationDto {
-
-    private String name;
-    private Double longitude;
-    private Double latitude;
-
-}
+//package com.example.omg_project.domain.trip.dto;
+//
+//import lombok.*;
+//
+//@Getter
+//@Setter
+//@NoArgsConstructor
+//@AllArgsConstructor
+//public class TripLocationDto {
+//
+//    private String name;
+//    private Double longitude;
+//    private Double latitude;
+//
+//}
