+++ src/main/java/com/example/omg_project/domain/trip/service/TripLocationService.java
@@ -1,13 +1,13 @@
-package com.example.omg_project.domain.trip.service;
-
-import com.example.omg_project.domain.trip.dto.TripLocationDto;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public interface TripLocationService {
-
-    void saveLocations(List<TripLocationDto> tripLocationDtoList);
-
-}
+//package com.example.omg_project.domain.trip.service;
+//
+//import com.example.omg_project.domain.trip.dto.TripLocationDTO;
+//import org.springframework.stereotype.Service;
+//
+//import java.util.List;
+//
+//@Service
+//public interface TripLocationService {
+//
+//    void saveLocations(List<TripLocationDTO> tripLocationDTOList);
+//
+//}
